Use a tags for slides link
The links to slides are http://www.benmvp.com/slides so Gatsby things they are pages w/in the site so it tries to be smart, but we really need to load a separate page. So by making them an `<a>` tag that opens in a new window so gatsby doesn't try to control things.